run for existing code

+++ tests/Hamcrest/Type/IsDoubleTest.php
@@ -12,7 +12,7 @@ class IsDoubleTest extends \Hamcrest\AbstractMatcherTest
     public function testEvaluatesToTrueIfArgumentMatchesType()
     {
         assertThat((float) 5.2, floatValue());
-        assertThat((double) 5.3, doubleValue());
+        assertThat((float) 5.3, doubleValue());
     }
 
     public function testEvaluatesToFalseIfArgumentDoesntMatchType()
